refactor Controls and AudioPlayer
